requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.4.4
Django==3.1.6
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.3
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
